+++ database/pgsql/vulnerability.go
@@ -26,6 +26,73 @@ import (
 	"github.com/coreos/clair/ext/versionfmt"
 )
 
+const (
+	lockVulnerabilityAffects = `LOCK vulnerability_affected_namespaced_feature IN SHARE ROW EXCLUSIVE MODE`
+
+	searchVulnerability = `
+		SELECT v.id, v.description, v.link, v.severity, v.metadata, n.version_format 
+		FROM vulnerability AS v, namespace AS n
+		WHERE v.namespace_id = n.id
+		AND v.name = $1
+		AND n.name = $2
+		AND v.deleted_at IS NULL
+		`
+
+	insertVulnerabilityAffected = `
+		INSERT INTO vulnerability_affected_feature(vulnerability_id, feature_name, affected_version, fixedin)
+		VALUES ($1, $2, $3, $4)
+		RETURNING ID
+	`
+
+	searchVulnerabilityAffected = `
+		SELECT vulnerability_id, feature_name, affected_version, fixedin 
+		FROM vulnerability_affected_feature
+		WHERE vulnerability_id = ANY($1)
+	`
+
+	searchVulnerabilityByID = `
+		SELECT v.name, v.description, v.link, v.severity, v.metadata, n.name, n.version_format
+		FROM vulnerability AS v, namespace AS n
+		WHERE v.namespace_id = n.id
+			AND v.id = $1`
+
+	searchVulnerabilityPotentialAffected = `
+		WITH req AS (
+			SELECT vaf.id AS vaf_id, n.id AS n_id, vaf.feature_name AS name, v.id AS vulnerability_id
+			FROM vulnerability_affected_feature AS vaf,
+				vulnerability AS v,
+				namespace AS n
+			WHERE vaf.vulnerability_id = ANY($1)
+			AND v.id = vaf.vulnerability_id
+			AND n.id = v.namespace_id
+			)
+		SELECT req.vulnerability_id, nf.id, f.version, req.vaf_id AS added_by
+		FROM feature AS f, namespaced_feature AS nf, req
+		WHERE f.name = req.name
+		AND nf.namespace_id = req.n_id
+		AND nf.feature_id = f.id`
+
+	insertVulnerabilityAffectedNamespacedFeature = `
+		INSERT INTO vulnerability_affected_namespaced_feature(vulnerability_id, namespaced_feature_id, added_by)
+		VALUES ($1, $2, $3)`
+
+	insertVulnerability = `
+		WITH ns AS (
+			SELECT id FROM namespace WHERE name = $6 AND version_format = $7
+		)
+		INSERT INTO Vulnerability(namespace_id, name, description, link, severity, metadata, created_at)
+		VALUES((SELECT id FROM ns), $1, $2, $3, $4, $5, CURRENT_TIMESTAMP)
+		RETURNING id`
+
+	removeVulnerability = `
+		UPDATE Vulnerability
+		SET deleted_at = CURRENT_TIMESTAMP
+		WHERE namespace_id = (SELECT id FROM Namespace WHERE name = $1)
+			AND name = $2
+			AND deleted_at IS NULL
+		RETURNING id`
+)
+
 var (
 	errVulnerabilityNotFound = errors.New("vulnerability is not in database")
 )
